handle larger time_t ok guenther

+++ usr.bin/ranlib/touch.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: touch.c,v 1.6 2003/06/12 20:58:10 deraadt Exp $	*/
+/*	$OpenBSD: touch.c,v 1.7 2009/10/27 23:59:42 deraadt Exp $	*/
 
 /*-
  * Copyright (c) 1990 The Regents of the University of California.
@@ -75,7 +75,7 @@ settime(int afd)
 	if (lseek(afd, size, SEEK_SET) == (off_t)-1)
 		error(archive);
 	(void)snprintf(buf, sizeof buf,
-	    "%-12ld", (long int)time((time_t *)NULL) + RANLIBSKEW);
+	    "%-12lld", (long long)time((time_t *)NULL) + RANLIBSKEW);
 	if (write(afd, buf, sizeof(hdr->ar_date)) != sizeof(hdr->ar_date))
 		error(archive);
 }
